tweaks

+++ homeassistant/components/fan/__init__.py
@@ -152,7 +152,7 @@ async def async_setup(hass, config: dict):
         SERVICE_INCREASE_SPEED,
         {
             vol.Optional(ATTR_PERCENTAGE_STEP): vol.All(
-                vol.Coerce(float), vol.Range(min=0, max=100)
+                vol.Coerce(int), vol.Range(min=0, max=100)
             )
         },
         "async_increase_speed",
@@ -162,7 +162,7 @@ async def async_setup(hass, config: dict):
         SERVICE_DECREASE_SPEED,
         {
             vol.Optional(ATTR_PERCENTAGE_STEP): vol.All(
-                vol.Coerce(float), vol.Range(min=0, max=100)
+                vol.Coerce(int), vol.Range(min=0, max=100)
             )
         },
         "async_decrease_speed",
